requirement.txt
Django==4.0.4
djangorestframework==3.15.1
boto3
django-storages
python-dotenv
djangorestframework-simplejwt
textract-trp
django-json-widget